// contracts/generated/typescript/models/CLIToken.ts
export interface CLIToken {
    /**
     *
     * @type {string}
     * @memberof CLIToken
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof CLIToken
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof CLIToken
     */
    tokenPrefix: string;
    /**
     *
     * @type {CLITokenStatusEnum}
     * @memberof CLIToken
     */
    status: CLITokenStatusEnum;
    /**
     *
     * @type {Date}
     * @memberof CLIToken
     */
    expiresAt: Date;
    /**
     *
     * @type {Date}
     * @memberof CLIToken
     */
    lastUsedAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof CLIToken
     */
    revokedAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof CLIToken
     */
    createdAt: Date;
}


/**
 * @export
 */
export const CLITokenStatusEnum = {
    Active: 'active',
    Expired: 'expired',
    Revoked: 'revoked'
} as const;
export type CLITokenStatusEnum = typeof CLITokenStatusEnum[keyof typeof CLITokenStatusEnum];